Less flawed benchmark for filtering.
The previous benchmark was not entirely fair as it was not accessing
any properties of the filtered documents.

This version accesses the ._key property of each element, but there is
no actual filtering happening. When doing real filtering the performance
should become aligned to the set() performance where the main bottleneck
is pushing elements onto the array.

Anyway, this being said, here are the new, less-flawed numbers:

105800 docs added in 1009ms
flushed to disc, starting filtering ...

Filtered 25180400 docs in 1004 ms 	(12540040 per sec)