Made filter benchmark more honest
Now there is actual filtering happening, and we're still getting
~10mio. docs / sec processed. Need to start testing with only filtering
through all records once, but that will take a lot of records for
testing. After that there might still be some v8 optimizations which
are screwing with the performance, but we're certainly getting closer
to the truth here.

975100 docs added in 10003ms
flushed to disc, starting filtering ...

Filtered 98485100 docs in 10066 ms 	(9783936 per sec)