Events+callbacks never fire when in test mode.
By passing a null/empty path to Dirty it will run in test mode.  This is
useful when re-running a process multiple times when testing and not
wanting to save state.  However, no callbacks or events are fired in
that mode, making it hard to actually do testing.  These changes will
make sure all async actions are triggered when in normal or test mode.